motor driver  v1.0.0

+++ Baremetal/ADASWithSDV/src/utils/utils.c
@@ -7,8 +7,8 @@
 #include <stdint.h>
 #include "stm32f4xx.h"
 #include "./utils.h"
-uint8_t PinSource(uint32_t pin) {
-    uint8_t pinSource = 0;
+uint16_t PinSource(uint32_t pin) {
+    uint16_t pinSource = 0;
     while (pin >>= 1  )
     	pinSource++;
     return pinSource;
